requirements.txt
flask>=3.0.0
gtts>=2.5.1
Flask-Compress>=1.13
python-dotenv>=1.0.1
requests>=2.32.3
openai>=1.35.10
